pyproject.toml
[tool.poetry]
name = "remnashop"
version = "2.0.0"
description = "This is a Telegram bot for selling VPN subscriptions. It works with Remnawave"
authors = ["snoups"]
readme = "README.md"

packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.20.0.post0"
aiogram-dialog = "^2.3.1"
alembic = "^1.15.2"
fastapi = "^0.115.12"
pydantic-settings = "^2.9.1"
redis = "^6.1.0"
uvicorn = "^0.34.2"
asyncpg = "^0.30.0"
celery = "^5.5.2"
aiogram-i18n = "^1.4"
fluent-runtime = "^0.4.0"
msgspec = "^0.19.0"
greenlet = "^3.2.2"
remnawave-api = "^1.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"